# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple manga reader app with rate, favorite and comment features",
        "title": "Manga Explorer API",
        "contact": {
            "name": "arcorium",
            "url": "github.com/arcorium",
            "email": "arcorium.l@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/devices": {
            "get": {
                "description": "Get all user credentials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get Credentials",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.CredentialResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/common.FieldError"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "login to authenticate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "user login input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "$ref": "#/definitions/dto.LoginResponse"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/common.FieldError"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logout/{cred_id}": {
            "post": {
                "description": "logout credential id or current logged in credential",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "$ref": "#/definitions/common.ParameterError"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logouts": {
            "post": {
                "description": "Remove all current user logged in credentials",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout All Device",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/common.FieldError"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/refresh-token": {
            "post": {
                "description": "Recreate access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh Token",
                "parameters": [
                    {
                        "description": "refresh token input, token type should be Bearer",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RefreshTokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "$ref": "#/definitions/dto.RefreshTokenResponse"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/common.FieldError"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.FieldError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        },
        "common.ParameterError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "param": {
                    "type": "string"
                }
            }
        },
...
        "dto.CredentialResponse": {
            "type": "object",
            "properties": {
                "device_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorWrapper": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorResponse"
                }
            }
        },
...
        "dto.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "dto.RefreshTokenInput": {
            "type": "object",
            "required": [
                "access_token",
                "token_type"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "dto.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
...
        "dto.ResponsePage": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "page": {
                    "$ref": "#/definitions/dto.ResponsePage"
                }
            }
        },
        "dto.SuccessWrapper": {
            "type": "object",
            "properties": {
                "success": {
                    "$ref": "#/definitions/dto.SuccessResponse"
                }
            }
        },
...
    }
}